+++ src/Math/Matrix4.elm
@@ -6,7 +6,6 @@ module Math.Matrix4
         , inverse
         , inverseOrthonormal
         , makeBasis
-        , makeFromList
         , makeFrustum
         , makeLookAt
         , makeOrtho
@@ -68,8 +67,8 @@ existing matrix.
 
 -}
 
+import Elm.Kernel.MJS
 import Math.Vector3 exposing (Vec3)
-import Native.MJS
 
 
 {-| 4x4 matrix type
@@ -82,14 +81,14 @@ type Mat4
 -}
 transform : Mat4 -> Vec3 -> Vec3
 transform =
-    Native.MJS.v3mul4x4
+    Elm.Kernel.MJS.v3mul4x4
 
 
 {-| A matrix with all 0s, except 1s on the diagonal.
 -}
 identity : Mat4
 identity =
-    Native.MJS.m4x4identity
+    Elm.Kernel.MJS.m4x4identity
 
 
 {-| Computes the inverse of any matrix. This is somewhat computationally
@@ -97,7 +96,7 @@ intensive. If the matrix is not invertible, `Nothing` is returned.
 -}
 inverse : Mat4 -> Maybe Mat4
 inverse =
-    Native.MJS.m4x4inverse
+    Elm.Kernel.MJS.m4x4inverse
 
 
 {-| Computes the inverse of the given matrix, assuming that the matrix is
@@ -106,7 +105,7 @@ has no possibility of failing.
 -}
 inverseOrthonormal : Mat4 -> Mat4
 inverseOrthonormal =
-    Native.MJS.m4x4inverseOrthonormal
+    Elm.Kernel.MJS.m4x4inverseOrthonormal
 
 
 {-| Creates a matrix for a projection frustum with the given parameters.
@@ -123,7 +122,7 @@ Parameters:
 -}
 makeFrustum : Float -> Float -> Float -> Float -> Float -> Float -> Mat4
 makeFrustum =
-    Native.MJS.m4x4makeFrustum
+    Elm.Kernel.MJS.m4x4makeFrustum
 
 
 {-| Creates a matrix for a perspective projection with the given parameters.
@@ -138,7 +137,7 @@ Parameters:
 -}
 makePerspective : Float -> Float -> Float -> Float -> Mat4
 makePerspective =
-    Native.MJS.m4x4makePerspective
+    Elm.Kernel.MJS.m4x4makePerspective
 
 
 {-| Creates a matrix for an orthogonal frustum projection with the given parameters.
@@ -155,7 +154,7 @@ Parameters:
 -}
 makeOrtho : Float -> Float -> Float -> Float -> Float -> Float -> Mat4
 makeOrtho =
-    Native.MJS.m4x4makeOrtho
+    Elm.Kernel.MJS.m4x4makeOrtho
 
 
 {-| Creates a matrix for a 2D orthogonal frustum projection with the given
@@ -171,21 +170,21 @@ Parameters:
 -}
 makeOrtho2D : Float -> Float -> Float -> Float -> Mat4
 makeOrtho2D =
-    Native.MJS.m4x4makeOrtho2D
+    Elm.Kernel.MJS.m4x4makeOrtho2D
 
 
 {-| Matrix multiplcation: a * b
 -}
 mul : Mat4 -> Mat4 -> Mat4
 mul =
-    Native.MJS.m4x4mul
+    Elm.Kernel.MJS.m4x4mul
 
 
 {-| Matrix multiplication, assuming a and b are affine: a * b
 -}
 mulAffine : Mat4 -> Mat4 -> Mat4
 mulAffine =
-    Native.MJS.m4x4mulAffine
+    Elm.Kernel.MJS.m4x4mulAffine
 
 
 {-| Creates a transformation matrix for rotation in radians about the
@@ -193,14 +192,14 @@ mulAffine =
 -}
 makeRotate : Float -> Vec3 -> Mat4
 makeRotate =
-    Native.MJS.m4x4makeRotate
+    Elm.Kernel.MJS.m4x4makeRotate
 
 
 {-| Concatenates a rotation in radians about an axis to the given matrix.
 -}
 rotate : Float -> Vec3 -> Mat4 -> Mat4
 rotate =
-    Native.MJS.m4x4rotate
+    Elm.Kernel.MJS.m4x4rotate
 
 
 {-| Creates a transformation matrix for scaling by 3 scalar values, one for
@@ -208,7 +207,7 @@ each of the x, y, and z directions.
 -}
 makeScale3 : Float -> Float -> Float -> Mat4
 makeScale3 =
-    Native.MJS.m4x4makeScale3
+    Elm.Kernel.MJS.m4x4makeScale3
 
 
 {-| Creates a transformation matrix for scaling each of the x, y, and z axes by
@@ -216,21 +215,21 @@ the amount given in the corresponding element of the 3-element vector.
 -}
 makeScale : Vec3 -> Mat4
 makeScale =
-    Native.MJS.m4x4makeScale
+    Elm.Kernel.MJS.m4x4makeScale
 
 
 {-| Concatenates a scaling to the given matrix.
 -}
 scale3 : Float -> Float -> Float -> Mat4 -> Mat4
 scale3 =
-    Native.MJS.m4x4scale3
+    Elm.Kernel.MJS.m4x4scale3
 
 
 {-| Concatenates a scaling to the given matrix.
 -}
 scale : Vec3 -> Mat4 -> Mat4
 scale =
-    Native.MJS.m4x4scale
+    Elm.Kernel.MJS.m4x4scale
 
 
 {-| Creates a transformation matrix for translating by 3 scalar values, one for
@@ -238,7 +237,7 @@ each of the x, y, and z directions.
 -}
 makeTranslate3 : Float -> Float -> Float -> Mat4
 makeTranslate3 =
-    Native.MJS.m4x4makeTranslate3
+    Elm.Kernel.MJS.m4x4makeTranslate3
 
 
 {-| Creates a transformation matrix for translating each of the x, y, and z
@@ -246,21 +245,21 @@ axes by the amount given in the corresponding element of the 3-element vector.
 -}
 makeTranslate : Vec3 -> Mat4
 makeTranslate =
-    Native.MJS.m4x4makeTranslate
+    Elm.Kernel.MJS.m4x4makeTranslate
 
 
 {-| Concatenates a translation to the given matrix.
 -}
 translate3 : Float -> Float -> Float -> Mat4 -> Mat4
 translate3 =
-    Native.MJS.m4x4translate3
+    Elm.Kernel.MJS.m4x4translate3
 
 
 {-| Concatenates a translation to the given matrix.
 -}
 translate : Vec3 -> Mat4 -> Mat4
 translate =
-    Native.MJS.m4x4translate
+    Elm.Kernel.MJS.m4x4translate
 
 
 {-| Creates a transformation matrix for a camera.
@@ -274,7 +273,7 @@ Parameters:
 -}
 makeLookAt : Vec3 -> Vec3 -> Vec3 -> Mat4
 makeLookAt =
-    Native.MJS.m4x4makeLookAt
+    Elm.Kernel.MJS.m4x4makeLookAt
 
 
 {-| "Flip" the matrix across the diagonal by swapping row index and column
@@ -282,33 +281,25 @@ index.
 -}
 transpose : Mat4 -> Mat4
 transpose =
-    Native.MJS.m4x4transpose
+    Elm.Kernel.MJS.m4x4transpose
 
 
 {-| Creates a transform from a basis consisting of 3 linearly independent vectors.
 -}
 makeBasis : Vec3 -> Vec3 -> Vec3 -> Mat4
 makeBasis =
-    Native.MJS.m4x4makeBasis
-
-
-{-| Creates a matrix from a list of elements. Returns Nothing if the length of
-the list is not exactly 16 (4x4).
--}
-makeFromList : List Float -> Maybe Mat4
-makeFromList =
-    Native.MJS.m4x4fromList
+    Elm.Kernel.MJS.m4x4makeBasis
 
 
 {-| Convert a matrix to a record.
 -}
 toRecord : Mat4 -> { m11 : Float, m21 : Float, m31 : Float, m41 : Float, m12 : Float, m22 : Float, m32 : Float, m42 : Float, m13 : Float, m23 : Float, m33 : Float, m43 : Float, m14 : Float, m24 : Float, m34 : Float, m44 : Float }
 toRecord =
-    Native.MJS.m4x4toRecord
+    Elm.Kernel.MJS.m4x4toRecord
 
 
 {-| Convert a record to a matrix.
 -}
 fromRecord : { m11 : Float, m21 : Float, m31 : Float, m41 : Float, m12 : Float, m22 : Float, m32 : Float, m42 : Float, m13 : Float, m23 : Float, m33 : Float, m43 : Float, m14 : Float, m24 : Float, m34 : Float, m44 : Float } -> Mat4
 fromRecord =
-    Native.MJS.m4x4fromRecord
+    Elm.Kernel.MJS.m4x4fromRecord
