+++ src/Math/Vector2.elm
@@ -51,7 +51,7 @@ The set functions create a new copy of the vector, updating a single field.
 
 -}
 
-import Native.Math.Vector2
+import Elm.Kernel.Math.Vector2
 
 
 {-| Two dimensional vector type
@@ -64,137 +64,137 @@ type Vec2
 -}
 vec2 : Float -> Float -> Vec2
 vec2 =
-    Native.Math.Vector2.vec2
+    Elm.Kernel.MJS.v2
 
 
 {-| Extract the x component of a vector.
 -}
 getX : Vec2 -> Float
 getX =
-    Native.Math.Vector2.getX
+    Elm.Kernel.MJS.v2getX
 
 
 {-| Extract the y component of a vector.
 -}
 getY : Vec2 -> Float
 getY =
-    Native.Math.Vector2.getY
+    Elm.Kernel.MJS.v2getY
 
 
 {-| Update the x component of a vector, returning a new vector.
 -}
 setX : Float -> Vec2 -> Vec2
 setX =
-    Native.Math.Vector2.setX
+    Elm.Kernel.MJS.v2setX
 
 
 {-| Update the y component of a vector, returning a new vector.
 -}
 setY : Float -> Vec2 -> Vec2
 setY =
-    Native.Math.Vector2.setY
+    Elm.Kernel.MJS.v2setY
 
 
 {-| Convert a vector to a tuple.
 -}
 toTuple : Vec2 -> ( Float, Float )
 toTuple =
-    Native.Math.Vector2.toTuple
+    Elm.Kernel.MJS.v2toTuple
 
 
 {-| Convert a vector to a record.
 -}
 toRecord : Vec2 -> { x : Float, y : Float }
 toRecord =
-    Native.Math.Vector2.toRecord
+    Elm.Kernel.MJS.v2toRecord
 
 
 {-| Convert a tuple to a vector.
 -}
 fromTuple : ( Float, Float ) -> Vec2
 fromTuple =
-    Native.Math.Vector2.fromTuple
+    Elm.Kernel.MJS.v2fromTuple
 
 
 {-| Convert a record to a vector.
 -}
 fromRecord : { x : Float, y : Float } -> Vec2
 fromRecord =
-    Native.Math.Vector2.fromRecord
+    Elm.Kernel.MJS.v2fromRecord
 
 
 {-| Vector addition: a + b
 -}
 add : Vec2 -> Vec2 -> Vec2
 add =
-    Native.Math.Vector2.add
+    Elm.Kernel.MJS.v2add
 
 
 {-| Vector subtraction: a - b
 -}
 sub : Vec2 -> Vec2 -> Vec2
 sub =
-    Native.Math.Vector2.sub
+    Elm.Kernel.MJS.v2sub
 
 
 {-| Vector negation: -a
 -}
 negate : Vec2 -> Vec2
 negate =
-    Native.Math.Vector2.neg
+    Elm.Kernel.MJS.v2negate
 
 
 {-| The normalized direction from b to a: (a - b) / |a - b|
 -}
 direction : Vec2 -> Vec2 -> Vec2
 direction =
-    Native.Math.Vector2.direction
+    Elm.Kernel.MJS.v2direction
 
 
 {-| The length of the given vector: |a|
 -}
 length : Vec2 -> Float
 length =
-    Native.Math.Vector2.length
+    Elm.Kernel.MJS.v2length
 
 
 {-| The square of the length of the given vector: |a| * |a|
 -}
 lengthSquared : Vec2 -> Float
 lengthSquared =
-    Native.Math.Vector2.lengthSquared
+    Elm.Kernel.MJS.v2lengthSquared
 
 
 {-| The distance between two vectors.
 -}
 distance : Vec2 -> Vec2 -> Float
 distance =
-    Native.Math.Vector2.distance
+    Elm.Kernel.MJS.v2distance
 
 
 {-| The square of the distance between two vectors.
 -}
 distanceSquared : Vec2 -> Vec2 -> Float
 distanceSquared =
-    Native.Math.Vector2.distanceSquared
+    Elm.Kernel.MJS.v2distanceSquared
 
 
 {-| A unit vector with the same direction as the given vector: a / |a|
 -}
 normalize : Vec2 -> Vec2
 normalize =
-    Native.Math.Vector2.normalize
+    Elm.Kernel.MJS.v2normalize
 
 
 {-| Multiply the vector by a scalar: s * v
 -}
 scale : Float -> Vec2 -> Vec2
 scale =
-    Native.Math.Vector2.scale
+    Elm.Kernel.MJS.v2scale
 
 
 {-| The dot product of a and b
 -}
 dot : Vec2 -> Vec2 -> Float
 dot =
-    Native.Math.Vector2.dot
+    Elm.Kernel.MJS.v2dot
