+++ tests/Tests.elm
@@ -107,15 +107,6 @@ suite =
                                     |> M4.scale3 (1 / 195) (1 / 195) (1 / 195)
                                     |> Just
                                 )
-                , test "makeFromList should be able create identity matrix" <|
-                    \_ ->
-                        M4.makeFromList [ 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1 ] |> Expect.equal (Just M4.identity)
-                , test "makeFromList should return Nothing for short list of elements" <|
-                    \_ ->
-                        M4.makeFromList [] |> Expect.equal Nothing
-                , test "makeFromList should return Nothing for too long list of elements" <|
-                    \_ ->
-                        M4.makeFromList [ 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1 ] |> Expect.equal Nothing
                 , test "fromRecord should be able to create the identity matrix" <|
                     \_ ->
                         M4.fromRecord { m11 = 1, m21 = 0, m31 = 0, m41 = 0, m12 = 0, m22 = 1, m32 = 0, m42 = 0, m13 = 0, m23 = 0, m33 = 1, m43 = 0, m14 = 0, m24 = 0, m34 = 0, m44 = 1 } |> Expect.equal M4.identity
@@ -125,16 +116,6 @@ suite =
                 , fuzz fuzz4x4 "fromRecord should be the opposite of toRecord" <|
                     \record4x4 ->
                         M4.toRecord (M4.fromRecord record4x4) |> Expect.equal record4x4
-                , test "matrix from list should not lose precision" <|
-                    \_ ->
-                        let
-                            matrix =
-                                M4.makeFromList
-                                    [ 0.0007645259938837921, 0, 0, 0, 0, 0.0007645259938837921, 0, 0, -178844.39532879056, -115444.72328862673, -23930.400791413915, -0.4996215217852649, 179114.76199136386, 115619.29560235815, 23930.038231557068, 0.5003784782148529 ]
-                        in
-                        matrix
-                            |> toString
-                            |> Expect.equal "Just { 0 = 0.0007645259938837921, 1 = 0, 2 = 0, 3 = 0, 4 = 0, 5 = 0.0007645259938837921, 6 = 0, 7 = 0, 8 = -178844.39532879056, 9 = -115444.72328862673, 10 = -23930.400791413915, 11 = -0.4996215217852649, 12 = 179114.76199136386, 13 = 115619.29560235815, 14 = 23930.038231557068, 15 = 0.5003784782148529 }"
                 , test "matrix make basis should not lose precision" <|
                     \_ ->
                         let
